[GH-3] refactor: Added nest-like service structure
