+++ server/src/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import ArtistsModule from './artists/artists.module';
+import { ArtistsModule } from './artists/artists.module';
 
 @Module({
   imports: [ArtistsModule],
