+++ server/src/artists/artists.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import ArtistsController from './artists.controller';
-import ArtistsService from './services';
+import ArtistsService from './services/artists.service';
+import WikiService from './services/wiki.service';
+import SpotifyService from './services/spotify.service';
 
 @Module({
   controllers: [ArtistsController],
-  providers: [ArtistsService],
+  providers: [ArtistsService, WikiService, SpotifyService],
 })
-export default class ArtistsModule {}
+export class ArtistsModule {}
